backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
llama-cpp-python==0.2.20
playwright==1.40.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
lxml==4.9.3